+++ Changelog.md
@@ -10,6 +10,7 @@
 - Fix: Inserting fuel into any type of passcode-protected furnace using hoppers/pipes/etc. puts the fuel into the wrong slot
 - Fix: Items other than colored lenses can be inserted into some lens slots using hoppers/pipes/etc.
 - Fix: More than one colored lens can be inserted into some lens slots using hoppers/pipes/etc.
+- Fix: The color of a Claymore/Trophy System does not update when inserting a lens using hoppers/pipes/etc.
 
 --------------------------Changelog for v1.9.8-beta1 of SecurityCraft--------------------------